Fix imports to build desktop client
