+++ src/settings/EditNotificationEmailDialog.ts
@@ -17,8 +17,8 @@ import { PayloadTooLargeError } from "../api/common/error/RestError"
 import { SegmentControl } from "../gui/base/SegmentControl"
 import { insertInlineImageB64ClickHandler } from "../mail/view/MailViewerUtils"
 import { UserError } from "../api/main/UserError"
-import { getAvailablePlansWithWhitelabel, showNotAvailableForFreeDialog, showPlanUpgradeRequiredDialog } from "../misc/SubscriptionDialogs"
-import { isWhitelabelActive } from "../subscription/SubscriptionUtils"
+import { showNotAvailableForFreeDialog, showPlanUpgradeRequiredDialog } from "../misc/SubscriptionDialogs"
+import { getAvailablePlansWithWhitelabel, isWhitelabelActive } from "../subscription/SubscriptionUtils"
 import type { UserController } from "../api/main/UserController"
 import { GENERATED_MAX_ID } from "../api/common/utils/EntityUtils"
 import { locator } from "../api/main/MainLocator"
