+++ src/templates/TemplateGroupUtils.ts
@@ -1,9 +1,10 @@
 import type { TemplateGroupRoot } from "../api/entities/tutanota/TypeRefs.js"
 import { TemplateGroupRootTypeRef } from "../api/entities/tutanota/TypeRefs.js"
-import { getAvailablePlansWithTemplates, showPlanUpgradeRequiredDialog } from "../misc/SubscriptionDialogs"
+import { showPlanUpgradeRequiredDialog } from "../misc/SubscriptionDialogs"
 import { locator } from "../api/main/MainLocator"
 import { FeatureType } from "../api/common/TutanotaConstants"
 import { isCustomizationEnabledForCustomer } from "../api/common/utils/Utils"
+import { getAvailablePlansWithTemplates } from "../subscription/SubscriptionUtils.js"
 
 /**
  * @return True if the group has been created.
