+++ test/tests/subscription/SubscriptionUtilsTest.ts
@@ -3,7 +3,8 @@ import { NewPaidPlans, PlanType } from "../../../src/api/common/TutanotaConstant
 import { IServiceExecutor } from "../../../src/api/common/ServiceRequest.js"
 import { createUpgradePriceServiceMock, PLAN_PRICES } from "./priceTestUtils.js"
 import { clone } from "@tutao/tutanota-utils"
-import { getAvailableMatchingPlans } from "../../../src/misc/SubscriptionDialogs.js"
+
+import { getAvailableMatchingPlans } from "../../../src/subscription/SubscriptionUtils.js"
 
 o.spec("SubscriptionUtilsTest", function () {
 	let serviceExecutor: IServiceExecutor
